+++ Example/Tests/BoundaryTests.swift
@@ -63,7 +63,7 @@ class BoundaryTests: XCTestCase {
         let featureConfiguration =
         """
         [{
-        "name": "Example A/B Test",
+        "name": "Example Feature A/B Test",
         "enabled": true,
         "test-variations": ["Enabled", "Disabled"]
         }]
@@ -84,7 +84,7 @@ class BoundaryTests: XCTestCase {
         let featureConfiguration =
         """
         [{
-        "name": "Example A/B Test",
+        "name": "Example Feature A/B Test",
         "enabled": true,
         "test-variations": ["Enabled", "Disabled"]
         }]
@@ -105,7 +105,7 @@ class BoundaryTests: XCTestCase {
         let featureConfiguration =
         """
         [{
-        "name": "Example A/B Test",
+        "name": "Example Feature A/B Test",
         "enabled": false,
         "test-variations": ["Enabled", "Disabled"]
         }]
@@ -120,6 +120,26 @@ class BoundaryTests: XCTestCase {
         feature.testVariationAssignment = 0.0
         XCTAssert(!feature.isTestVariation(.enabled), "Pass")
         XCTAssert(feature.isTestVariation(.disabled), "Pass")
+    }
+
+    func testWhenFeatureABNotEnabledThatIsDisabled() {
+        let featureConfiguration =
+        """
+        [{
+        "name": "Example Feature A/B Test",
+        "enabled": false,
+        "test-variations": ["Enabled", "Disabled"]
+        }]
+        """
+        let decoder = JSONDecoder()
+        guard let featuresData = featureConfiguration.data(using: .utf8),
+            let features = try? decoder.decode([Feature].self, from: featuresData),
+            var feature = features.first else {
+                XCTAssert(false, "Fail")
+                return
+        }
+        feature.testVariationAssignment = 0.0
+        XCTAssert(!feature.isEnabled(), "Pass")
     }
 
     func testWhenTestVariationAssignmentIsSmallAndNonZeroThatVariationIsGroupA() {