+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.3.0] - 2018-12-06
 ### Added
 - Added `FeatureFlagsUI.autoRefresh` property to automatically refresh data from configuration when the app is foregrounded (defaults to `false`).
 - Added ability to clear cache from `FeatureFlagsViewController`.
+- Added convenience initializer to `Feature` allowing a Feature to be retrieved using `Feature(named: .myFeature)` rather than `Feature.named(.myFeature)`.
 
 ## [1.2.2] - 2018-12-05
 ### Added